+++ projects/services-layers/src/lib/layers.service.spec.ts
@@ -1,12 +1,23 @@
 import { TestBed, async } from '@angular/core/testing';
 import { RasterLayer, Layer, CustomLayer, TGeoExtent } from './types/Layers';
+import { WmsLayer, WmtsLayer } from './types/RasterLayers';
 import { LayerGroup } from './types/LayerGroup';
 import { LayersService } from './layers.service';
 import { first } from 'rxjs/operators';
 import { BehaviorSubject } from 'rxjs';
 
 
-let layer1: Layer, layer2: Layer, layergroup1: LayerGroup, layergroup2: LayerGroup, layer3: Layer, layer4: Layer, layer5: CustomLayer;
+let layer1: Layer,
+  layer2: Layer,
+  layergroup1: LayerGroup,
+  layergroup2: LayerGroup,
+  layer3: Layer,
+  layer4: Layer,
+  layer5: CustomLayer,
+  layergroup3: LayerGroup,
+  layer6: WmsLayer,
+  layer7: WmtsLayer;
+
 
 
 describe('LayersService', () => {
@@ -38,10 +49,10 @@ describe('LayersService', () => {
 
     layergroup1 = new LayerGroup({
       visible: false,
-      name: 'baslayer Group',
+      name: 'baselayer Group',
       filtertype: 'Overlays',
-      id: 'baselayer_group',
-      layers: [layer1]
+      id: 'layergroup1',
+      layers: []
     });
 
     layer3 = new RasterLayer({
@@ -69,9 +80,10 @@ describe('LayersService', () => {
     layergroup2 = new LayerGroup({
       name: 'Group 2',
       filtertype: 'Overlays',
-      id: 'group2',
-      layers: [layer1, layer4]
+      id: 'layergroup2',
+      layers: []
     });
+
 
     layer5 = new CustomLayer({
       id: 'ID-vector-image',
@@ -81,14 +93,65 @@ describe('LayersService', () => {
       visible: false
     });
 
+    layergroup3 = new LayerGroup({
+      visible: true,
+      name: 'Group 3',
+      filtertype: 'Layers',
+      id: 'group3',
+      layers: []
+    });
+
+    layer6 = new WmsLayer({
+      name: 'test layer6',
+      type: 'wms',
+      id: 'layer6',
+      url: '//geoservice.dlr.de/eoc/basemap/wms',
+      removable: true,
+      params: {
+        LAYERS: 'litemap'
+      }
+    });
+
+    layer7 = new WmtsLayer({
+      name: 'test layer7',
+      type: 'wmts',
+      id: 'layer7',
+      url: 'https://tiles.geoservice.dlr.de/service/wmts',
+      removable: true,
+      params: {
+        layer: 'eoc:world_relief_bw',
+        style: '_empty',
+        matrixSetOptions: {
+          matrixSet: 'EPSG:3857',
+          tileMatrixPrefix: 'EPSG:3857'
+        }
+      }
+    });
+
+    const service: LayersService = TestBed.get(LayersService);
+    service.setLayerGroups([]);
   });
-
-
 
   it('should be created', () => {
     const service: LayersService = TestBed.get(LayersService);
     expect(service).toBeTruthy();
   });
+
+  it('should update a Subject', async(() => {
+    const subject1 = new BehaviorSubject([layer5]);
+    expect(subject1.getValue()[0]).toBe(layer5);
+
+    const newLayer = new CustomLayer({
+      id: 'ID-vector-image',
+      name: 'Custom Layer KML',
+      type: 'custom',
+      custom_layer: new Object({ ol_uid: 600 }),
+      visible: true
+    });
+
+    subject1.next([newLayer]);
+    expect(subject1.getValue()[0]).toBe(newLayer);
+  }));
 
   /**
    * logs - 'layer or Group with id: layer1 already exists!
@@ -96,32 +159,37 @@ describe('LayersService', () => {
   it('should not add a layer if the id already exists!', async(() => {
     const service: LayersService = TestBed.get(LayersService);
     service.addLayer(layer1, 'Layers');
+    expect(service.getLayersCount()).toEqual(1);
+
     service.addLayer(layer1, 'Layers');
     expect(service.getLayersCount()).toEqual(1);
-    // clean up
-    service.removeLayer(layer1);
-    expect(service.getLayersCount()).toEqual(0);
   }));
 
   it('should add a layer to BaseLayers', async(() => {
     const service: LayersService = TestBed.get(LayersService);
     service.addLayer(layer1, 'Baselayers');
     expect(service.getBaseLayersCount()).toEqual(1);
-    // clean up
-    service.removeLayer(layer1);
-    expect(service.getBaseLayersCount()).toEqual(0);
   }));
 
   it('should remove a layer from BaseLayers', async(() => {
     const service: LayersService = TestBed.get(LayersService);
     service.addLayer(layer1, 'Baselayers');
-    expect(service.getBaseLayersCount()).toEqual(1);
+    service.addLayer(layer2, 'Baselayers');
+    expect(service.getBaseLayersCount()).toEqual(2);
 
     service.removeLayer(layer1, 'Baselayers');
-    expect(service.getBaseLayersCount()).toEqual(0);
+    expect(service.getBaseLayersCount()).toEqual(1);
   }));
 
-  it('should remove a layer with its filtertype even it was set on add Layer', async(() => {
+
+  it('should set the filtertype on a layer by adding it to the store', async(() => {
+    const service: LayersService = TestBed.get(LayersService);
+    service.addLayer(layer1, 'Overlays');
+    const _layer1 = service.getLayerById('layer1');
+    expect(_layer1.filtertype).toEqual('Overlays');
+  }));
+
+  it('should remove a layer and use its filtertype', async(() => {
     const service: LayersService = TestBed.get(LayersService);
     expect(layer1.filtertype).toEqual('Layers');
     service.addLayer(layer1, 'Overlays');
@@ -137,7 +205,7 @@ describe('LayersService', () => {
     const service: LayersService = TestBed.get(LayersService);
     service.addLayer(layer1, 'Overlays');
     expect(service.getOverlaysCount()).toEqual(1);
-    service.removeLayer(layer1, 'Overlays');
+    service.removeLayerOrGroupById('layer1');
     expect(service.getOverlaysCount()).toEqual(0);
   }));
 
@@ -145,9 +213,6 @@ describe('LayersService', () => {
     const service: LayersService = TestBed.get(LayersService);
     service.addLayer(layer1, 'Layers');
     expect(service.getLayersCount()).toEqual(1);
-    // clean up
-    service.removeLayer(layer1, layer1.filtertype);
-    expect(service.getLayersCount()).toEqual(0);
   }));
 
   it('should remove a layer from Layers', async(() => {
@@ -160,118 +225,115 @@ describe('LayersService', () => {
     expect(service.getLayersCount()).toEqual(0);
   }));
 
-  it('should add a layer to Overlays', async(() => {
-    const service: LayersService = TestBed.get(LayersService);
-    service.addLayer(layer1, 'Overlays');
-    expect(service.getOverlaysCount()).toEqual(1);
-    // clean up
-    service.removeLayer(layer1, layer1.filtertype);
-    expect(service.getOverlaysCount()).toEqual(0);
-  }));
-
-  it('should remove a layer from Overlays', async(() => {
+  it('should add/remove a layer to Overlays', async(() => {
     const service: LayersService = TestBed.get(LayersService);
     service.addLayer(layer1, 'Overlays');
     expect(service.getOverlaysCount()).toEqual(1);
 
     service.removeLayer(layer1, 'Overlays');
     expect(service.getOverlaysCount()).toEqual(0);
-
-    // clean up;
-    service.setLayerGroups([]);
-    expect(service.getLayerGroupsCount()).toEqual(0);
   }));
 
   it('should add a layer to a layerGroup', async(() => {
     const service: LayersService = TestBed.get(LayersService);
-    // layergroup1.filtertype = 'Layers';
+    layergroup1.layers.push(layer1);
     service.addLayerGroup(layergroup1, 'Layers');
     expect(service.getLayersCount()).toEqual(1);
     expect(service.getLayerGroupsCount()).toEqual(1);
 
     service.addLayerToGroup(layer2, layergroup1);
     expect(service.getLayersCount()).toEqual(2);
-    // clean up
-    service.removeLayerGroup(layergroup1);
-    expect(service.getLayerGroupsCount()).toEqual(0);
-    expect(service.getLayersCount()).toEqual(0);
   }));
 
 
   it('should add a layerGroup to Layers', async(() => {
     const service: LayersService = TestBed.get(LayersService);
-    // layergroup1.filtertype = 'Layers';
+    layergroup1.layers.push(layer1);
     service.addLayerGroup(layergroup1, 'Layers');
     expect(service.getLayersCount()).toEqual(1);
     expect(service.getLayerGroupsCount()).toEqual(1);
-    // clean up
-    service.removeLayerGroup(layergroup1);
-    expect(service.getLayerGroupsCount()).toEqual(0);
-    expect(service.getLayersCount()).toEqual(0);
   }));
 
   it('should add a layerGroup and replace its filtertype on itself and all layers', async(() => {
     const service: LayersService = TestBed.get(LayersService);
-    // layergroup1.filtertype = 'Layers';
+    layergroup1.layers.push(layer1);
+    expect(layer1.filtertype).toEqual('Layers');
     expect(layergroup1.filtertype).toEqual('Overlays');
-    service.addLayerGroup(layergroup1, 'Layers');
-    expect(service.getLayersCount()).toEqual(1);
+    const replaceFiltertype = 'Baselayers';
+
+    service.addLayerGroup(layergroup1, replaceFiltertype);
+    expect(service.getBaseLayersCount()).toEqual(1);
     expect(service.getLayerGroupsCount()).toEqual(1);
-    expect(layergroup1.filtertype).toEqual('Layers');
+    expect(layergroup1.filtertype).toEqual(replaceFiltertype);
     // check if filtertype is replaced on all Layers
     layergroup1.layers.map(l => {
-      expect(l.filtertype).toEqual('Layers');
+      expect(l.filtertype).toEqual(replaceFiltertype);
     });
-    // clean up
-    service.removeLayerGroup(layergroup1);
-    expect(service.getLayerGroupsCount()).toEqual(0);
-    expect(service.getLayersCount()).toEqual(0);
-    expect(service.getOverlaysCount()).toEqual(0);
   }));
 
   it('should remove a layerGroup by ID', async(() => {
     const service: LayersService = TestBed.get(LayersService);
     layergroup1.id = 'removable_group';
+    layergroup1.layers.push(layer1);
     service.addLayerGroup(layergroup1, 'Overlays');
     service.removeLayerOrGroupById('removable_group');
     expect(service.getLayerGroupsCount()).toEqual(0);
+  }));
 
-    // clean up;
-    service.setLayerGroups([]);
+  // this creates - LOG: 'layerGroup: layergroup1 is not removable!'
+  it('should not remove a non removable layer', async(() => {
+    const service: LayersService = TestBed.get(LayersService);
+    layer1.removable = false;
+    service.addLayer(layer1, 'Baselayers');
+    service.removeLayer(layer1);
+    expect(service.getLayerGroupsCount()).toEqual(1);
+  }));
+
+  // this creates - LOG: 'layer: layer1 is removed with force!'
+  it('should remove a non removable layer with force', async(() => {
+    const service: LayersService = TestBed.get(LayersService);
+    layer1.removable = false;
+    service.addLayer(layer1, 'Baselayers');
+    service.removeLayer(layer1, layer1.filtertype, true);
     expect(service.getLayerGroupsCount()).toEqual(0);
   }));
 
-  // this creates - LOG: 'layerGroup: baselayer_group is not removable!'
+  // this creates - LOG: 'layerGroup: layergroup1 is not removable!'
   it('should not remove a non removable layerGroup', async(() => {
     const service: LayersService = TestBed.get(LayersService);
+    layergroup1.layers.push(layer1);
     layergroup1.removable = false;
     service.addLayerGroup(layergroup1, 'Baselayers');
     service.removeLayerGroup(layergroup1);
     expect(service.getLayerGroupsCount()).toEqual(1);
+  }));
 
-    // clean up;
-    service.setLayerGroups([]);
+  // this creates - LOG: 'layerGroup: layergroup1 is not removable!'
+  it('should remove a non removable layerGroup with force', async(() => {
+    const service: LayersService = TestBed.get(LayersService);
+    layergroup1.layers.push(layer1);
+    layergroup1.removable = false;
+    service.addLayerGroup(layergroup1, 'Baselayers');
+    service.removeLayerGroup(layergroup1, true);
     expect(service.getLayerGroupsCount()).toEqual(0);
   }));
 
   it('should add a layerGroup to Overlays', async(() => {
     const service: LayersService = TestBed.get(LayersService);
-    // layergroup1.filtertype = 'Overlays';
+    layergroup1.layers.push(layer1);
     service.addLayerGroup(layergroup1, 'Overlays');
     expect(service.getOverlaysCount()).toEqual(1);
     expect(service.getLayerGroupsCount()).toEqual(1);
-    // clean up
-    service.removeLayerGroup(layergroup1);
-    expect(service.getLayerGroupsCount()).toEqual(0);
-    expect(service.getOverlaysCount()).toEqual(0);
   }));
 
   it('should remove a layer from a layerGroup but not the group', async(() => {
     const service: LayersService = TestBed.get(LayersService);
+    layergroup1.layers.push(layer1);
+    // layergroup1.removable = false;
     service.addLayerGroup(layergroup1, 'Overlays');
     expect(service.getLayerGroupsCount()).toEqual(1);
     expect(service.getOverlaysCount()).toEqual(1);
-    // clean up
+
     service.removeLayerFromGroup(layer1, layergroup1, false);
     expect(service.getLayerGroupsCount()).toEqual(1);
     expect(service.getOverlaysCount()).toEqual(0);
@@ -279,29 +341,32 @@ describe('LayersService', () => {
 
   it('should remove a layer by Id from a layerGroup and the group if only one layer was on the group', async(() => {
     const service: LayersService = TestBed.get(LayersService);
+    layergroup1.layers.push(layer1);
+    layergroup1.removable = true;
     service.addLayerGroup(layergroup1, 'Overlays');
     expect(service.getLayerGroupsCount()).toEqual(1);
     expect(service.getOverlaysCount()).toEqual(1);
-    // clean up;
-    service.removeLayerGroup(layergroup1);
+
+    service.removeLayerOrGroupById('layer1');
     expect(service.getLayerGroupsCount()).toEqual(0);
     expect(service.getOverlaysCount()).toEqual(0);
   }));
 
   it('should remove multiple or all layers from the Overlays-Slot', async(() => {
     const service: LayersService = TestBed.get(LayersService);
+    layergroup1.layers = [layer1, layer3];
     service.addLayerGroup(layergroup1, 'Overlays');
     service.addLayer(layer2, 'Overlays');
     // Layer and LayerGroups Count -getLayerGroupsCount()
     expect(service.getLayerGroupsCount()).toEqual(2);
-    expect(service.getOverlaysCount()).toEqual(2);
-    // clean up;
+    expect(service.getOverlaysCount()).toEqual(3);
+
     service.removeOverlays();
     expect(service.getLayerGroupsCount()).toEqual(0);
     expect(service.getOverlaysCount()).toEqual(0);
   }));
 
-  it('should remove multiple or all layers from the Layers-Slot', async(() => {
+  it('should remove multiple or all layers from the Layers-Slot with a filter', async(() => {
     const service: LayersService = TestBed.get(LayersService);
     [layer1, layer2, layer3].forEach(l => service.addLayer(l, 'Layers'));
     expect(service.getLayersCount()).toEqual(3);
@@ -310,65 +375,55 @@ describe('LayersService', () => {
       return layer.id !== layer1.id;
     });
     expect(service.getLayersCount()).toEqual(1);
-
-
-    // clean up;
-    service.setLayerGroups([]);
-    expect(service.getLayerGroupsCount()).toEqual(0);
   }));
 
-  it('should remove multiple or all layers from the BaseLayers-Slot', async(() => {
+  it('should remove multiple or all layers from the BaseLayers-Slot with a filter', async(() => {
     const service: LayersService = TestBed.get(LayersService);
-    [layer1, layer2, layer3].forEach(l => service.addLayer(l, 'Baselayers'))
+    [layer1, layer2, layer3].forEach(l => service.addLayer(l, 'Baselayers'));
     expect(service.getBaseLayersCount()).toEqual(3);
-    // clean up;
+
     service.removeBaseLayers((layer) => {
       return layer.id !== layer1.id;
     });
     expect(service.getBaseLayersCount()).toEqual(1);
-
-
-    // clean up;
-    service.setLayerGroups([]);
-    expect(service.getLayerGroupsCount()).toEqual(0);
   }));
 
-  it('should replace all Layers and Groups in LayerGroups', async(() => {
+  it('should get a Layer or Group by Id', async(() => {
     const service: LayersService = TestBed.get(LayersService);
-    const groups = [layergroup1, layer2, layergroup2];
-    service.setLayerGroups(groups);
-    expect(service.getLayerOrGroupById('baselayer_group')).toBeDefined();
-    expect(service.getLayerOrGroupById('layer2')).toBeDefined();
-    expect(service.getLayerOrGroupById('group2')).toBeDefined();
+    layergroup1.layers.push(layer1);
+    service.addLayerGroup(layergroup1, 'Overlays');
 
-    // clean up;
-    service.setLayerGroups([]);
-    expect(service.getLayerGroupsCount()).toEqual(0);
+    expect(service.getOverlaysCount()).toBe(1);
+    expect(service.getLayerGroupsCount()).toBe(1);
+
+    const _layer1 = service.getLayerOrGroupById('layer1');
+    expect(_layer1).toBe(layer1);
+    const _layergroup1 = service.getLayerOrGroupById('layergroup1');
+    expect(_layergroup1).toBe(layergroup1);
   }));
 
   it('should get all LayerGroups', async(() => {
     const service: LayersService = TestBed.get(LayersService);
+    layergroup1.layers.push(layer1);
+    layergroup2.layers = [layer1, layer4];
     const groups = [layergroup1, layer2, layergroup2];
     service.setLayerGroups(groups);
+
     service.getLayerGroups().pipe(first()).subscribe(_groups => {
       expect(_groups.length).toEqual(3);
-
-      // clean up;
-      service.setLayerGroups([]);
-      expect(service.getLayerGroupsCount()).toEqual(0);
     });
   }));
 
   it('should get all Overlays', async(() => {
     const service: LayersService = TestBed.get(LayersService);
+    layergroup2.layers = [layer1, layer4];
     service.addLayerGroup(layergroup2, 'Overlays');
+    service.addLayer(layer3, 'Layers');
 
     service.getOverlays().pipe(first()).subscribe(overlays => {
       expect(overlays.length).toEqual(2);
-
-      // clean up;
-      service.setLayerGroups([]);
-      expect(service.getLayerGroupsCount()).toEqual(0);
+      expect(overlays[0]).toBe(layer1);
+      expect(overlays[1]).toBe(layer4);
     });
   }));
 
@@ -378,10 +433,6 @@ describe('LayersService', () => {
 
     service.getLayers().pipe(first()).subscribe(layers => {
       expect(layers.length).toEqual(2);
-
-      // clean up;
-      service.setLayerGroups([]);
-      expect(service.getLayerGroupsCount()).toEqual(0);
     });
   }));
 
@@ -391,31 +442,7 @@ describe('LayersService', () => {
 
     service.getBaseLayers().pipe(first()).subscribe(baselayers => {
       expect(baselayers.length).toEqual(2);
-
-      // clean up;
-      service.setLayerGroups([]);
-      expect(service.getLayerGroupsCount()).toEqual(0);
     });
-  }));
-
-
-  it('should update a Subject', async(() => {
-    const subject1 = new BehaviorSubject([layer5]);
-    expect(subject1.getValue()[0]).toBe(layer5);
-
-
-    const newLayer = new CustomLayer({
-      id: 'ID-vector-image',
-      name: 'Custom Layer KML',
-      type: 'custom',
-      custom_layer: new Object({ ol_uid: 600 }),
-      visible: true
-    });
-
-    subject1.next([newLayer]);
-    expect(subject1.getValue()[0]).toBe(newLayer);
-
-
   }));
 
   it('should update a Layer', async(() => {
@@ -436,16 +463,13 @@ describe('LayersService', () => {
 
     expect(layerFromSvc.visible).toEqual(true);
     expect((layerFromSvc as CustomLayer).custom_layer.ol_uid).toEqual(600);
-
-    // clean up;
-    service.setLayerGroups([]);
-    expect(service.getLayerGroupsCount()).toEqual(0);
   }));
 
 
   it('should update a LayerGroup', async(() => {
     const service: LayersService = TestBed.get(LayersService);
-
+    layer1.visible = false;
+    layergroup1.layers.push(layer1);
     service.addLayerGroup(layergroup1);
     expect(layergroup1.visible).toBe(false);
     expect(layergroup1.bbox).toBe(undefined); // like defined above
@@ -457,47 +481,49 @@ describe('LayersService', () => {
       visible: true,
       name: 'baslayer Group',
       filtertype: 'Layers',
-      id: 'baselayer_group',
+      id: 'layergroup1',
       bbox: bbox,
-      layers: [layer5]
+      layers: [layer5, layer2]
     });
     // set by group
     expect(layer5.visible).toBe(newLayergroup1.visible);
 
     service.updateLayerGroup(newLayergroup1);
-    const layerFromSvc = service.getLayerOrGroupById(newLayergroup1.id)[0];
-    console.log(layerFromSvc);
+    const layerFromSvc = service.getLayerOrGroupById(newLayergroup1.id);
 
     expect(layerFromSvc.visible).toEqual(newLayergroup1.visible);
     expect((layerFromSvc as LayerGroup).layers[0]).toEqual(layer5);
+    expect((layerFromSvc as LayerGroup).layers[1]).toEqual(layer2);
     expect((layerFromSvc as LayerGroup).bbox).toEqual(bbox);
-
-    // clean up;
-    service.setLayerGroups([]);
-    expect(service.getLayerGroupsCount()).toEqual(0);
   }));
 
   // test for layer-control arrows up and down
   it('should move group or layer up and down in the array', async(() => {
     const service: LayersService = TestBed.get(LayersService);
-    // layer2 == 0, layer1 == 1, layer4 == 2 layer3 == 3
+    layergroup2.layers = [layer1, layer4];
     const layers = [layer2, layergroup2, layer3];
+    // layer2,       layer1, layer4,      layer3
     service.setLayerGroups(layers);
 
-    // layer2 after layer1 and layer4
     service.setGroupLayerIndex(layer2, 'down');
-    // layer4 before layer1
+    // layer2 after layer1 and layer4
+    service.getLayerGroups().pipe(first()).subscribe(groups => {
+      const flatt = service.flattenDeepArray(groups);
+      expect(flatt[0]).toBe(layer1);
+      expect(flatt[1]).toBe(layer4);
+      expect(flatt[2]).toBe(layer2);
+      expect(flatt[3]).toBe(layer3);
+    });
+
     service.setLayerIndexInGroup(layer4, 'up', layergroup2);
+    // layer4 before layer1
+
     service.getLayerGroups().pipe(first()).subscribe(groups => {
       const flatt = service.flattenDeepArray(groups);
       expect(flatt[0]).toBe(layer4);
       expect(flatt[1]).toBe(layer1);
       expect(flatt[2]).toBe(layer2);
       expect(flatt[3]).toBe(layer3);
-
-      // clean up;
-      service.setLayerGroups([]);
-      expect(service.getLayerGroupsCount()).toEqual(0);
     });
   }));
 
@@ -505,28 +531,116 @@ describe('LayersService', () => {
   it('should set correct zIndex for groups or layers', async(() => {
     const service: LayersService = TestBed.get(LayersService);
     // layer2 == 0, layer1 == 1, layer4 == 2 layer3 == 3
+    layergroup2.layers = [layer1, layer4];
     const layers = [layer2, layergroup2, layer3];
     service.setLayerGroups(layers);
 
-    // layer2 after layer1 and layer4
+
     service.arrayMove(layers, 1, 0);
     service.setLayerGroups(layers);
-    // layer4 before layer1
+    // layer2 after layer1 and layer4
+    service.getLayerGroups().pipe(first()).subscribe(groups => {
+      const flatt = service.flattenDeepArray(groups);
+      expect(flatt[0]).toBe(layer1);
+      expect(flatt[1]).toBe(layer4);
+      expect(flatt[2]).toBe(layer2);
+      expect(flatt[3]).toBe(layer3);
+    });
+
+
     service.arrayMove(layergroup2.layers, 1, 0);
     service.updateLayerGroup(layergroup2);
+    // layer4 before layer1
+
     service.getLayerGroups().pipe(first()).subscribe(groups => {
       const flatt = service.flattenDeepArray(groups);
       expect(flatt[0]).toBe(layer4);
       expect(flatt[1]).toBe(layer1);
       expect(flatt[2]).toBe(layer2);
       expect(flatt[3]).toBe(layer3);
+    });
+  }));
 
-      // clean up;
-      service.setLayerGroups([]);
-      expect(service.getLayerGroupsCount()).toEqual(0);
+  it('should replace all Layers and Groups in LayerGroups', async(() => {
+    const service: LayersService = TestBed.get(LayersService);
+    layergroup1.layers.push(layer3);
+    const groups1 = [layergroup1];
+    service.setLayerGroups(groups1);
+
+    service.getLayerGroups().pipe(first()).subscribe(groups => {
+      const flatt = service.flattenDeepArray(groups);
+      expect(flatt.length).toBe(1);
+      expect(flatt[0]).toBe(layer3);
+    });
+
+
+    layergroup2.layers = [layer1, layer4];
+    const groups2 = [layergroup1, layer2, layergroup2];
+    service.setLayerGroups(groups2);
+    // replace all Groups
+
+    service.getLayerGroups().pipe(first()).subscribe(groups => {
+      const flatt = service.flattenDeepArray(groups);
+      expect(flatt.length).toBe(4);
+      expect(flatt[0]).toBe(layer3);
+      expect(flatt[1]).toBe(layer2);
+      expect(flatt[2]).toBe(layer1);
+      expect(flatt[3]).toBe(layer4);
     });
   }));
 
 
+  it('should set/reset all Groups or layers for a filtertype', async(() => {
+    const service: LayersService = TestBed.get(LayersService);
+    // add only Baselayers -----------------------
+    const baseLayers = [layer1, layer2];
+    service.setLayerGroups(baseLayers, 'Baselayers');
+
+    expect(service.getBaseLayersCount()).toBe(2);
+    service.getLayerGroups().pipe(first()).subscribe(groups => {
+      const flatt = service.flattenDeepArray(groups);
+      expect(flatt.length).toBe(2);
+      expect(flatt[0]).toBe(layer1);
+      expect(flatt[1]).toBe(layer2);
+    });
+
+    // add Layers above Baselayers ----------------
+    layergroup3.layers = [layer3, layer4, layer5];
+    const layers = [layergroup3, layer6, layer7];
+    service.setLayerGroups(layers, 'Layers');
+
+
+    expect(service.getLayersCount()).toBe(5);
+
+    service.getLayerGroups().pipe(first()).subscribe(groups => {
+      const flatt = service.flattenDeepArray(groups);
+      expect(flatt.length).toBe(7);
+      expect(flatt[0]).toBe(layer1);
+      expect(flatt[1]).toBe(layer2);
+
+      expect(flatt[2]).toBe(layer3);
+      expect(flatt[3]).toBe(layer4);
+      expect(flatt[4]).toBe(layer5);
+      expect(flatt[5]).toBe(layer6);
+      expect(flatt[6]).toBe(layer7);
+    });
+
+
+    // reset Layers -------------------------------
+    const layersNew = [layer6, layer7];
+    service.setLayerGroups(layersNew, 'Layers');
+    expect(service.getLayersCount()).toBe(2);
+
+    service.getLayerGroups().pipe(first()).subscribe(groups => {
+      const flatt = service.flattenDeepArray(groups);
+      expect(flatt.length).toBe(4);
+
+      expect(flatt[0]).toBe(layer1);
+      expect(flatt[1]).toBe(layer2);
+
+      expect(flatt[2]).toBe(layer6);
+      expect(flatt[3]).toBe(layer7);
+    });
+  }));
 
 });